Resolved 66 merge conflicts. Core upstream changes in this release:

- DOM: MutationEffectOnScript enum added to CharacterData mutation API;
  nsTextFragment renamed to CharacterDataBuffer (mText -> mBuffer)
- DOM: Location::SetURI renamed to Navigate
- SpiderMonkey: stateLoop inlined into nsHtml5Tokenizer header with SIMD
  dispatch (not adopted -- Foxhound keeps monolithic stateLoop with taint)
- SpiderMonkey: str_split refactored to use std::find(); taint re-integrated
- SpiderMonkey: moz.build reorganised -- proxy/threading/vm files moved to
  subdirectory builds
- Networking: InstallCacheListener refactored to delegate to
  DoInstallCacheListener; URL length checks added to nsStandardURL
- js/loader: LoadedScript SRI/bytecode storage refactored
  (mScriptBytecode -> mSRIAndSerializedStencil)
- config: check_spidermonkey_style.py updated for new ICU4X/Glean headers

Foxhound taint tracking preserved throughout all changed subsystems.

+++ js/src/frontend/EmitterScope.h
@@ -57,7 +57,7 @@ class MOZ_STACK_CLASS EmitterScope : public Nestable<EmitterScope> {
 #endif
 
   // The number of enclosing environments. Used for error checking.
-  uint8_t environmentChainLength_;
+  uint16_t environmentChainLength_;
 
   // The next usable slot on the frame for not-closed over bindings.
   //
